+++ packages/ledger-icp/src/utils/accounts.utils.spec.ts
@@ -1,5 +1,5 @@
 import { InvalidAccountIDError } from "../errors/ledger.errors";
-import { checkAccountId } from "./accounts.utils";
+import { checkAccountId, isIcpAccountIdentifier } from "./accounts.utils";
 
 describe("accounts-utils", () => {
   describe("checkAccountId", () => {
@@ -12,6 +12,24 @@ describe("accounts-utils", () => {
     it("should throw if not valid account id", () => {
       const call1 = () => checkAccountId("not-valid");
       expect(call1).toThrow(InvalidAccountIDError);
+    });
+  });
+
+  describe("isIcpAccountIdentifier", () => {
+    it("should return false if input is undefined", () => {
+      expect(isIcpAccountIdentifier(undefined)).toBe(false);
+    });
+
+    it("should return true if valid account id", () => {
+      expect(
+        isIcpAccountIdentifier(
+          "cd70bfa0f092c38a0ff8643d4617219761eb61d199b15418c0b1114d59e30f8e",
+        ),
+      ).toBe(true);
+    });
+
+    it("should return false if not valid account id", () => {
+      expect(isIcpAccountIdentifier("not-valid")).toBe(false);
     });
   });
 });