+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Features
 
+- Add utility `isIcpAccountIdentifier` to check if a string is a valid ICP account identifier.
+- Utility `notEmptyString` to use a type predicate, ensuring that the input is a string type when it returns true.
 - Support `disburseMaturity` in `@dfinity/nns`.
 
 # 2025.03.10-1330Z